+++ Tests/Wolverine.AspNetCore.Controllers/Wolverine.AspNetCore.Controllers.Application/Products/GetProductById/GetProductByIdQueryHandler.cs
@@ -21,12 +21,12 @@ namespace Wolverine.AspNetCore.Controllers.Application.Products.GetProductById
         }
 
         [IntentManaged(Mode.Merge, Signature = Mode.Fully, Body = Mode.Fully)]
-        public async Task<ProductDto> Handle(GetProductByIdQuery query, CancellationToken cancellationToken)
+        public async Task<ProductDto> Handle(GetProductByIdQuery request, CancellationToken cancellationToken)
         {
-            var product = await _productRepository.FindByIdAsync(query.Id, cancellationToken);
+            var product = await _productRepository.FindByIdAsync(request.Id, cancellationToken);
             if (product is null)
             {
-                throw new NotFoundException($"Could not find Product '{query.Id}'");
+                throw new NotFoundException($"Could not find Product '{request.Id}'");
             }
             return product.MapToProductDto(_mapper);
         }