+++ Tests/Wolverine.Mapperly/Wolverine.Mapperly.Application/Products/GetProductById/GetProductByIdHandler.cs
@@ -22,12 +22,12 @@ namespace Wolverine.Mapperly.Application.Products.GetProductById
         }
 
         [IntentManaged(Mode.Merge, Signature = Mode.Fully, Body = Mode.Fully)]
-        public async Task<ProductDto> Handle(GetProductById query, CancellationToken cancellationToken)
+        public async Task<ProductDto> Handle(GetProductById request, CancellationToken cancellationToken)
         {
-            var product = await _productRepository.FindByIdAsync(query.Id, cancellationToken);
+            var product = await _productRepository.FindByIdAsync(request.Id, cancellationToken);
             if (product is null)
             {
-                throw new NotFoundException($"Could not find Product '{query.Id}'");
+                throw new NotFoundException($"Could not find Product '{request.Id}'");
             }
             return _mapper.ProductToProductDto(product);
         }